requirements.txt
aplus==0.11.0
argon2-cffi==20.1.0
astropy==4.2.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
bqplot==0.12.25
branca==0.4.2
Brotli==1.0.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cryptography==3.4.6
cycler==0.10.0
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
dask==2021.4.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
FinanceDatabase==0.1.7
Flask==1.1.2
Flask-Compress==1.9.0
Flask-SQLAlchemy==2.5.1
frozendict==1.2
fsspec==0.9.0
future==0.18.2
greenlet==1.0.0
h5py==3.2.1
idna==2.10
ipydatawidgets==4.2.0
ipykernel==5.5.0
ipyleaflet==0.13.6
ipympl==0.7.0
ipython==7.21.0
ipython-genutils==0.2.0
ipyvolume==0.5.2
ipyvue==1.5.0
ipyvuetify==1.6.2
ipywebrtc==0.6.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
llvmlite==0.36.0
locket==0.2.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
multitasking==0.0.9
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
newsapi-python==0.2.6
notebook==6.3.0
numba==0.53.1
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
progressbar2==3.53.1
prometheus-client==0.10.1
prompt-toolkit==3.0.17
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==3.0.0
pycparser==2.20
pyerfa==1.7.2
Pygments==2.8.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-utils==2.5.6
pythreejs==2.3.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
SQLAlchemy==1.4.9
tabulate==0.8.9
td-ameritrade-python-api==0.3.4
terminado==0.9.4
testpath==0.4.4
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
traittypes==0.2.1
urllib3==1.26.4
vaex==4.1.0
vaex-astro==0.8.0
vaex-core==4.1.0
vaex-hdf5==0.7.0
vaex-jupyter==0.6.0
vaex-ml==0.11.1
vaex-server==0.4.0
vaex-viz==0.5.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xarray==0.17.0
yfinance==0.1.59
# dash is required to call `build:py`
dash[dev]>=1.15.0
